// types/enum/SupermarketLogos.ts
export enum SupermarketLogos {
  Mercadona = "../../assets/images/supermarkets/mercadona.png",
  Continente = "../../assets/images/supermarkets/continente.png",
  Lidl = "../../assets/images/supermarkets/lidl.png",
  PingoDoce = "../../assets/images/supermarkets/pingodoce.png",
  Intermarche = "../../assets/images/supermarkets/intermarche.png",
  Auchan = "../../assets/images/supermarkets/auchan.png",
  Minipreco = "../../assets/images/supermarkets/minipreco.png",
}